hw1/requirements.txt
PyYAML==5.4.1
requests==2.22.0
click==7.1.2